Refactor file name CombatLogReader.py to dungeon_mapper.py
